Moved deprecated syntax that is staying in 2.0 to a new file to keep it clear from the real file #400
